Diretório ViewModels

+++ SAT.MODELS/ViewModels/ViewDashboardDisponibilidadeTecnicosMediaGlobal.cs
@@ -4,6 +4,6 @@ namespace SAT.MODELS.ViewModels
     public class ViewDashboardDisponibilidadeTecnicosMediaGlobal
     {
         public string Classificacao { get; set; }
-        public int? Media { get; set; }
+        public double? Media { get; set; }
     }
 }
